+++ .gitmodules
@@ -1,3 +1,3 @@
-[submodule "src/app/core/Magento/Paymentwall/Controller/Index/paymentwall-php"]
-	path = src/app/core/Magento/Paymentwall/Controller/Index/paymentwall-php
+[submodule "src/lib/internal/paymentwall-php"]
+	path = src/lib/internal/paymentwall-php
 	url = git@github.com:paymentwall/paymentwall-php.git
