+++ src/app/code/Paymentwall/Paymentwall/etc/config.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Store:etc/config.xsd">
+    <default>
+        <payment>
+            <paymentwall>
+                <active>1</active>
+                <model>PaymentwallFacade</model>
+                <order_status>pending_payment</order_status>
+                <payment_action>authorize</payment_action>
+                <title>Paymentwall</title>
+                <currency>USD</currency>
+                <can_use_checkout>1</can_use_checkout>
+                <is_gateway>1</is_gateway>
+                <sort_order>1</sort_order>
+            </paymentwall>
+        </payment>
+    </default>
+</config>
