+++ src/app/code/Paymentwall/Paymentwall/etc/di.xml
@@ -1,0 +1,45 @@
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
+    <!-- Payment Method Facade configuration -->
+    <virtualType name="PaymentwallFacade" type="Magento\Payment\Model\Method\Adapter">
+        <arguments>
+            <argument name="code" xsi:type="const">\Paymentwall\Paymentwall\Model\Ui\ConfigProvider::CODE</argument>
+            <argument name="formBlockType" xsi:type="string">Magento\Payment\Block\Form</argument>
+            <argument name="infoBlockType" xsi:type="string">Paymentwall\Paymentwall\Block\Paymentwall</argument>
+            <argument name="valueHandlerPool" xsi:type="object">PaymentwallValueHandlerPool</argument>
+        </arguments>
+    </virtualType>
+
+    <!-- Configuration reader -->
+    <virtualType name="PaymentwallConfig" type="Magento\Payment\Gateway\Config\Config">
+        <arguments>
+            <argument name="methodCode" xsi:type="const">\Paymentwall\Paymentwall\Model\Ui\ConfigProvider::CODE</argument>
+        </arguments>
+    </virtualType>
+
+    <virtualType name="PaymentwallLogger" type="Magento\Payment\Model\Method\Logger">
+        <arguments>
+            <argument name="config" xsi:type="object">PaymentwallConfig</argument>
+        </arguments>
+    </virtualType> 
+ 
+    <!-- Value handlers infrastructure -->
+    <virtualType name="PaymentwallValueHandlerPool" type="Magento\Payment\Gateway\Config\ValueHandlerPool">
+        <arguments>
+            <argument name="handlers" xsi:type="array">
+                <item name="default" xsi:type="string">PaymentwallConfigValueHandler</item>
+            </argument>
+        </arguments>
+    </virtualType>
+    <virtualType name="PaymentwallConfigValueHandler" type="Magento\Payment\Gateway\Config\ConfigValueHandler">
+        <arguments>
+            <argument name="configInterface" xsi:type="object">PaymentwallConfig</argument>
+        </arguments>
+    </virtualType>
+
+    <type name="Paymentwall\Paymentwall\Block\Paymentwall">
+        <arguments>
+            <argument name="config" xsi:type="object">PaymentwallConfig</argument>
+        </arguments>
+    </type>
+
+</config>
