+++ src/app/code/Paymentwall/Paymentwall/etc/frontend/routes.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../../../../lib/internal/Magento/Framework/App/etc/routes.xsd">
+    <router id="standard">
+        <route id="paymentwall" frontName="paymentwall">
+            <module name="Paymentwall_Paymentwall" />
+        </route>
+    </router>
+</config>
