+++ src/app/code/Paymentwall/Paymentwall/registration.php
@@ -1,0 +1,7 @@
+<?php
+
+\Magento\Framework\Component\ComponentRegistrar::register(
+    \Magento\Framework\Component\ComponentRegistrar::MODULE,
+    'Paymentwall_Paymentwall',
+    __DIR__
+);
