+++ src/app/code/Paymentwall/Paymentwall/view/frontend/layout/paymentwall_index_index.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0"?>
+<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" layout="1column" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
+	<body>
+		<referenceContainer name="content">
+			<block class="Paymentwall\Paymentwall\Block\Paymentwall" name="paymentwall_paymentwall" template="paymentwall.phtml">
+			</block>
+		</referenceContainer>
+	</body>
+</page>
+
