--- src/app/core/Magento/Paymentwall/Model/Adminhtml/Source/PaymentAction.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace Magento\Paymentwall\Model\Adminhtml\Source;
- 
-/**
- * Class PaymentAction
- */
-class PaymentAction implements \Magento\Framework\Option\ArrayInterface
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function toOptionArray()
-    {
-        return [
-            [
-                'value' => AbstractMethod::ACTION_AUTHORIZE,
-                'label' => __('Authorize')
-            ]
-        ];
-    }
-}
