--- src/app/core/Magento/Paymentwall/etc/frontend/routes.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0"?>
- 
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../../../../../../lib/internal/Magento/Framework/App/etc/routes.xsd">
-    <router id="standard">
-        <route id="paymentwall" frontName="paymentwall">
-            <module name="Magento_Paymentwall" />
-        </route>
-    </router>
-</config>
