--- src/app/core/Magento/Paymentwall/registration.php
@@ -1,7 +1,0 @@
-<?php
-
-\Magento\Framework\Component\ComponentRegistrar::register(
-    \Magento\Framework\Component\ComponentRegistrar::MODULE,
-    'Magento_Paymentwall',
-    __DIR__
-);
