--- src/app/core/Magento/Paymentwall/view/frontend/layout/paymentwall_index_pingback.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0"?>
-<page xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" layout="1column" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
-    <body>
-        <referenceContainer name="content"></referenceContainer>
-    </body>
-</page>
