--- src/app/core/Magento/Paymentwall/view/frontend/web/js/view/payment/paymentwall.js
@@ -1,20 +1,0 @@
-define(
-    [
-        'uiComponent',
-        'Magento_Checkout/js/model/payment/renderer-list'
-    ],
-    function (
-        Component,
-        rendererList
-    ) {
-        'use strict';
-        rendererList.push(
-            {
-                type: 'paymentwall',
-                component: 'Magento_Paymentwall/js/view/payment/method-renderer/paymentwall'
-            }
-        );
-        /** Add view logic here if needed */
-        return Component.extend({});
-    }
-);
